refactor(pages): fix scroll flashes
