+++ src/components/nameDivider/nameDivider.svelte
@@ -3,7 +3,7 @@
 </script>
 
 <div
-  class="flex bg-orange-500 rounded-sm  h-14 justify-start items-center my-4 dark:text-white sticky z-20"
+  class="flex bg-orange-500 rounded-sm  h-14 justify-start items-center my-4 text-white sticky z-20"
 >
   <div class="w-1/6 h-1 rounded-sm " />
   <div class="w-4/6 text-center">
