+++ src/lib/index.ts
@@ -2,3 +2,4 @@ export * from './date'
 export * from './getOS'
 export * from './grid'
 export * from './isImage';
+export * from './search';
